chore; 뷰에 있던 비즈니스 로직들 뷰모델로 이동
